authorize permissions update+delete event

+++ controllers/eventController.js
@@ -1,9 +1,14 @@
 import { StatusCodes } from 'http-status-codes'
-import { NotFoundError } from '../errors/customErrors.js'
+import {
+  BadRequestError,
+  NotFoundError,
+  UnauthorizedError,
+} from '../errors/customErrors.js'
 import { PrismaClient } from '@prisma/client'
 const prisma = new PrismaClient()
 
 export const getAllEvents = async (req, res) => {
+  // console.log(req.user)
   const events = await prisma.events.findMany()
   res.status(StatusCodes.OK).json({ events })
 }
@@ -11,6 +16,8 @@ export const getAllEvents = async (req, res) => {
 export const createEvent = async (req, res) => {
   const { title, description, date, venueId, eventStatus, eventCategory } =
     req.body
+  // console.log(req.user)
+  req.body.createdById = req.user.userId
 
   const venue = await prisma.venues.findUnique({ where: { id: venueId } })
   if (!venue) throw new NotFoundError(`No venue with id ${venueId}`)
@@ -24,6 +31,7 @@ export const createEvent = async (req, res) => {
       venueId: venueIdDb,
       eventStatus,
       eventCategory,
+      createdById: req.body.createdById,
     },
   })
   res.status(StatusCodes.CREATED).json({ event })
@@ -47,6 +55,10 @@ export const updateEvent = async (req, res) => {
 
   const event = await prisma.events.findUnique({ where: { id } })
   if (!event) throw new NotFoundError(`No event with id ${id}`)
+
+  if (event.createdById !== req.user.userId && req.user.role[0] != 'ADMIN')
+    throw new UnauthorizedError('Not authorized to update this event')
+
   const updatedEvent = await prisma.events.update({
     where: { id },
     data: {
@@ -65,8 +77,15 @@ export const updateEvent = async (req, res) => {
 
 export const deleteEvent = async (req, res) => {
   const { id } = req.params
-  const event = await prisma.events.findUnique({ where: { id } })
+
+  const event = await prisma.events.findUnique({
+    where: { id },
+  })
   if (!event) throw new NotFoundError(`No event with id ${id}`)
+
+  if (event.createdById !== req.user.userId && req.user.role[0] != 'ADMIN')
+    throw new UnauthorizedError('Not authorized to delete this event')
+
   const deletedEvent = await prisma.events.delete({ where: { id } })
   res
     .status(StatusCodes.OK)
